refactor: refactor to adhere clean code structure
