+++ app/src/main/java/com/entsh118/housespot/ui/homepage/HomePageActivity.kt
@@ -1,4 +1,4 @@
-package com.entsh118.housespot.layananjasa
+package com.entsh118.housespot.ui.homepage
 
 import android.os.Bundle
 import androidx.activity.enableEdgeToEdge
@@ -7,11 +7,11 @@ import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
 import com.entsh118.housespot.R
 
-class FormPesananActivity : AppCompatActivity() {
+class HomePageActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
-        setContentView(R.layout.activity_form_pesanan)
+        setContentView(R.layout.activity_home_page)
         ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main)) { v, insets ->
             val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
             v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom)
