+++ app/src/main/java/com/entsh118/housespot/ui/layananjasa/DetailJasaActivity.kt
@@ -1,4 +1,4 @@
-package com.entsh118.housespot.pesanan
+package com.entsh118.housespot.ui.layananjasa
 
 import android.os.Bundle
 import androidx.activity.enableEdgeToEdge
@@ -7,11 +7,11 @@ import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
 import com.entsh118.housespot.R
 
-class PesananClientActivity : AppCompatActivity() {
+class DetailJasaActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
-        setContentView(R.layout.activity_pesanan_client)
+        setContentView(R.layout.activity_detail_jasa)
         ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main)) { v, insets ->
             val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
             v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom)
