+++ app/src/main/java/com/entsh118/housespot/ui/pesanan/PesananClientActivity.kt
@@ -1,4 +1,4 @@
-package com.entsh118.housespot.layananjasa
+package com.entsh118.housespot.ui.pesanan
 
 import android.os.Bundle
 import androidx.activity.enableEdgeToEdge
@@ -7,11 +7,11 @@ import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
 import com.entsh118.housespot.R
 
-class DetailJasaActivity : AppCompatActivity() {
+class PesananClientActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
-        setContentView(R.layout.activity_detail_jasa)
+        setContentView(R.layout.activity_pesanan_client)
         ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main)) { v, insets ->
             val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
             v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom)
