+++ app/src/main/res/layout/activity_account_homepage.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".account.AccountHomepageActivity">
+    tools:context=".ui.account.AccountHomepageActivity">
 
     <View
         android:id="@+id/rectangle_4"
