+++ app/src/main/res/layout/activity_detail_jasa.xml
@@ -4,7 +4,7 @@ xmlns:app="http://schemas.android.com/apk/res-auto"
 xmlns:tools="http://schemas.android.com/tools"
 android:layout_width="match_parent"
 android:layout_height="match_parent"
-tools:context=".layananjasa.DetailJasaActivity">
+tools:context=".ui.layananjasa.DetailJasaActivity">
 
     <ImageView
         android:id="@+id/detailPhoto"
