+++ app/src/main/res/layout/activity_form_pesanan.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".layananjasa.FormPesananActivity">
+    tools:context=".ui.layananjasa.FormPesananActivity">
 
     <RelativeLayout
         android:id="@+id/star"
