+++ app/src/main/res/layout/activity_home_page.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".homepage.HomePageActivity">
+    tools:context=".ui.homepage.HomePageActivity">
 
     <View
         android:id="@+id/rectangle_4"
