+++ app/src/main/res/layout/activity_pesanan_client.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".pesanan.PesananClientActivity">
+    tools:context=".ui.pesanan.PesananClientActivity">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/rvJasa"
